+++ src/main/java/com/beat/domain/admin/api/AdminController.java
@@ -1,10 +1,10 @@
 package com.beat.domain.admin.api;
 
-import com.beat.domain.admin.application.dto.CarouselFindAllResponse;
+import com.beat.domain.admin.application.dto.response.CarouselFindAllResponse;
 import com.beat.domain.admin.application.dto.request.CarouselProcessRequest;
 import com.beat.domain.admin.application.dto.response.CarouselProcessAllResponse;
 import com.beat.domain.admin.exception.AdminSuccessCode;
-import com.beat.domain.admin.application.dto.UserFindAllResponse;
+import com.beat.domain.admin.application.dto.response.UserFindAllResponse;
 import com.beat.domain.admin.facade.AdminFacade;
 import com.beat.global.auth.annotation.CurrentMember;
 import com.beat.global.common.dto.SuccessResponse;
