+++ src/main/java/com/beat/domain/admin/application/dto/response/UserFindAllResponse.java
@@ -1,4 +1,4 @@
-package com.beat.domain.admin.application.dto;
+package com.beat.domain.admin.application.dto.response;
 
 import com.beat.domain.user.domain.Users;
